Separate tracker life cycle and update tests
